+++ fabric/src/main/java/io/github/mortuusars/horseman/fabric/mixin/mount_gui/GuiMixin.java
@@ -1,6 +1,8 @@
 package io.github.mortuusars.horseman.fabric.mixin.mount_gui;
 
 import com.llamalad7.mixinextras.injector.ModifyReturnValue;
+import com.llamalad7.mixinextras.injector.wrapoperation.Operation;
+import com.llamalad7.mixinextras.injector.wrapoperation.WrapOperation;
 import io.github.mortuusars.horseman.Config;
 import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.Gui;
@@ -13,7 +15,6 @@ import org.spongepowered.asm.mixin.Shadow;
 import org.spongepowered.asm.mixin.Unique;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.ModifyVariable;
-import org.spongepowered.asm.mixin.injection.Redirect;
 
 @Mixin(value = Gui.class)
 public abstract class GuiMixin {
@@ -27,12 +28,31 @@ public abstract class GuiMixin {
     @Shadow
     protected abstract int getVehicleMaxHearts(LivingEntity vehicle);
 
-    @Shadow protected abstract boolean isExperienceBarVisible();
-
     @Unique
     private long horseman$lastTickVehicleInWater = -1;
 
-    @Redirect(method = "renderHotbarAndDecorations", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/player/LocalPlayer;jumpableVehicle()Lnet/minecraft/world/entity/PlayerRideableJumping;"))
+    @WrapOperation(method = "renderHotbarAndDecorations", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/player/LocalPlayer;jumpableVehicle()Lnet/minecraft/world/entity/PlayerRideableJumping;"))
+    private PlayerRideableJumping renderHotbarAndDecorations_jumpableVehicle(LocalPlayer player, Operation<PlayerRideableJumping> original) {
+        @Nullable PlayerRideableJumping vehicle = original.call(player);
+        if (vehicle == null || !Config.Client.IMPROVED_MOUNT_GUI.get()) return vehicle;
+
+        Minecraft mc = Minecraft.getInstance();
+
+        if (mc.gameMode == null || !mc.gameMode.hasExperience()) return vehicle;
+
+        if (vehicle instanceof LivingEntity entity && entity.isInWater()) {
+            horseman$lastTickVehicleInWater = player.level().getGameTime();
+        }
+
+        if (!mc.options.keyJump.isDown() && player.getJumpRidingScale() <= 0
+                || (player.level().getGameTime() - horseman$lastTickVehicleInWater < 10)) {
+            return null; // Prevent jump bar from rendering.
+        }
+
+        return vehicle;
+    }
+
+    /*@Redirect(method = "renderHotbarAndDecorations", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/player/LocalPlayer;jumpableVehicle()Lnet/minecraft/world/entity/PlayerRideableJumping;"))
     private PlayerRideableJumping renderHotbarAndDecorations_jumpableVehicle(LocalPlayer player) {
         @Nullable PlayerRideableJumping vehicle = player.jumpableVehicle();
 
@@ -52,17 +72,17 @@ public abstract class GuiMixin {
         }
 
         return vehicle;
-    }
+    }*/
 
-    @Redirect(method = "renderPlayerHealth", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;getVehicleMaxHearts(Lnet/minecraft/world/entity/LivingEntity;)I"))
-    private int renderPlayerHealth_getVehicleMaxHearts(Gui instance, LivingEntity entity) {
-        if (!Config.Client.IMPROVED_MOUNT_GUI.get()) return getVehicleMaxHearts(entity);
+    @WrapOperation(method = "renderPlayerHealth", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;getVehicleMaxHearts(Lnet/minecraft/world/entity/LivingEntity;)I"))
+    private int renderPlayerHealth_getVehicleMaxHearts(Gui instance, LivingEntity vehicle, Operation<Integer> original) {
+        if (!Config.Client.IMPROVED_MOUNT_GUI.get()) return original.call(instance, vehicle);
         return 0; // Forces hunger bar rendering, because it is not rendered when vehicle hearts is not 0.
     }
 
-    @Redirect(method = "renderPlayerHealth", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;getVisibleVehicleHeartRows(I)I"))
-    private int renderPlayerHealth_getVisibleVehicleHeartRows(Gui instance, int hearts) {
-        if (!Config.Client.IMPROVED_MOUNT_GUI.get()) return getVisibleVehicleHeartRows(hearts);
+    @WrapOperation(method = "renderPlayerHealth", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;getVisibleVehicleHeartRows(I)I"))
+    private int renderPlayerHealth_getVisibleVehicleHeartRows(Gui instance, int vehicleHealth, Operation<Integer> original) {
+        if (!Config.Client.IMPROVED_MOUNT_GUI.get()) return original.call(instance, vehicleHealth);
         // 'hears' will be 0 here, due to it being set in 'renderPlayerHealth_getVehicleMaxHearts'.
         LivingEntity livingEntity = getPlayerVehicleWithHealth();
         int vehicleHearts = getVehicleMaxHearts(livingEntity);
@@ -91,9 +111,9 @@ public abstract class GuiMixin {
         return original;
     }
 
-    @Redirect(method = "renderExperienceLevel", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;isExperienceBarVisible()Z"))
-    private boolean renderExperienceLevel(Gui instance) {
-        if (!Config.Client.IMPROVED_MOUNT_GUI.get()) return isExperienceBarVisible();
+    @WrapOperation(method = "renderExperienceLevel", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/Gui;isExperienceBarVisible()Z"))
+    private boolean renderExperienceLevel(Gui instance, Operation<Boolean> original) {
+        if (!Config.Client.IMPROVED_MOUNT_GUI.get()) return original.call(instance);
         // Always render exp level:
         return Minecraft.getInstance().gameMode != null && Minecraft.getInstance().gameMode.hasExperience();
     }